requirements.txt
anthropic>=0.40.0
tavily-python>=0.3.0
rich>=13.7.0
pydantic>=2.5.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.1.0
python-docx>=1.1.0
python-pptx>=0.6.23
extract-msg>=0.48.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
